Simplify ARM JIT packaging and tests
