+++ wie-arm-jit-types/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "wie-arm-jit"
+name = "wie-arm-jit-types"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
