+++ wie-arm-wasm/Cargo.toml
@@ -6,13 +6,10 @@ license.workspace = true
 description.workspace = true
 repository.workspace = true
 
-[features]
-compiler-runtime = []
-
 [dependencies]
 serde_json = { workspace = true }
 wasm-encoder = { version = "^0.257", default-features = false }
-wie-arm-jit = { workspace = true }
+wie-arm-jit-types = { workspace = true }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 js-sys = { version = "^0.3", default-features = false }
