--- wie-arm-wasm/compiler/.gitignore
@@ -1,2 +1,0 @@
-/pkg/
-/target/
