+++ wie-arm-wasm/src/compiler.rs
@@ -2,8 +2,9 @@ use alloc::string::ToString;
 
 use js_sys::{Object, Reflect, Uint8Array};
 use wasm_bindgen::prelude::*;
-use wie_arm_jit::CompileRequest;
+use wie_arm_jit_types::CompileRequest;
 
+#[wasm_bindgen]
 pub fn compile_request(payload: &[u8]) -> Result<JsValue, JsValue> {
     let request: CompileRequest = serde_json::from_slice(payload).map_err(|error| JsValue::from_str(&error.to_string()))?;
     let artifact = crate::codegen::compile(&request).map_err(|error| JsValue::from_str(&error))?;
