+++ wie-arm-wasm/src/lib.rs
@@ -1,16 +1,11 @@
+#![cfg(target_arch = "wasm32")]
 #![no_std]
 extern crate alloc;
 
 mod codegen;
 
-#[cfg(all(target_arch = "wasm32", feature = "compiler-runtime"))]
-pub mod compiler;
-#[cfg(all(target_arch = "wasm32", not(feature = "compiler-runtime")))]
+mod compiler;
 mod runtime;
 
 pub use codegen::{WasmArtifact, compile};
-#[cfg(all(target_arch = "wasm32", not(feature = "compiler-runtime")))]
 pub use runtime::WasmExecutor;
-
-#[cfg(test)]
-extern crate std;
