--- wie-arm-wasm/tests/browser-entry.js
@@ -1,4 +1,0 @@
-import { Probe } from "./browser-host/pkg/wie_arm_browser_test.js";
-import * as wasm from "./browser-host/pkg/wie_arm_browser_test_bg.wasm";
-
-globalThis.backendTest = { Probe, wasm };
