--- wie-arm-wasm/tests/browser-host/.gitignore
@@ -1,3 +1,0 @@
-/dist/
-/pkg/
-/target/
