+++ wie-web/src/ts/arm-compiler-worker.ts
@@ -1,8 +1,8 @@
-import("@wie-arm-compiler").then(({ compile_request }) => {
-    self.onmessage = async ({ data: { request, payload } }) => {
+import("@pkg").then(({ compile_request }) => {
+    self.onmessage = async ({ data: { request, payload } }: MessageEvent<{ request: string; payload: Uint8Array }>) => {
         try {
             if (!(payload instanceof Uint8Array)) throw new TypeError("compiler payload must be bytes");
-            const artifact = compile_request(payload);
+            const artifact: { bytes: Uint8Array<ArrayBuffer>; manifest: string } = compile_request(payload);
             const module = await WebAssembly.compile(artifact.bytes);
             self.postMessage({ request, module, manifest: artifact.manifest, encodedSize: artifact.bytes.length });
         } catch (error) {
